Enable ifunc attribute by default for ARM GNU/Linux
Similar to other architectures with IFUNC binutils/glibc support, this
patch enables the ifunc attribute for ARM GNU/Linux.  Although not
required for build master GLIBC, the intention is to allow refactor
its assembly implementation to C.

Tested compilation of glibc (in conjunction with a glibc patch to
support using the attribute on ARM) with build-many-glibcs.py (with
a patch to add a armv7 variant which enables multiarch).  I have
not run the GCC tests for ARM.

	* config.gcc (default_gnu_indirect_function): Default to yes for
	arm*-*-linux* with glibc.



git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@253635 138bc75d-0d04-0410-961f-82ee72b054a4